tests: adds tests associated with #9237
